# docs/v3/api/openapi.yaml
openapi: 3.1.0
info:
  title: Gotham Recon - Full Infrastructure API
  description: |
    Complete API documentation for the Gotham Recon security reconnaissance platform.

    ## Architecture Overview
    - **Graph Service** (8001): CQRS Read/Write for Asset Graph
    - **BFF Gateway** (8080): GraphQL API Gateway with real-time subscriptions
    - **Recon Orchestrator** (8000): Mission coordination and workflow management
    - **OSINT Runner** (8002): Phase 1 - Passive reconnaissance
    - **Active Recon** (8003): Phase 3 - HTTP probing and crawling
    - **Endpoint Intel** (8004): Phase 4 - Endpoint enrichment
    - **Verification** (8005): Phase 5 - Security testing
    - **Reporter** (8006): Phase 6 - Report generation
    - **Planner** (8007): Attack path scoring
    - **Scanner Proxy** (8051): gRPC interface for security tools

    ## Real-time Communication
    - WebSocket: Graph updates, Mission logs
    - SSE: Event streaming
    - GraphQL Subscriptions: Real-time data updates
  version: 1.0.0
  contact:
    name: Gotham Recon Team
  license:
    name: MIT
servers:
  - url: http://localhost:8001
    description: Graph Service
  - url: http://localhost:8080
    description: BFF Gateway
  - url: http://localhost:8000
    description: Recon Orchestrator
  - url: http://localhost:8002
    description: OSINT Runner
  - url: http://localhost:8003
    description: Active Recon
  - url: http://localhost:8004
    description: Endpoint Intel
  - url: http://localhost:8005
...
paths:
...
  /api/v1/nodes:
    post:
      tags: [Graph Service]
      summary: Create single node
      description: Create a new node in the asset graph
      operationId: createNode
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NodeCreate'
      responses:
        '201':
          description: Node created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NodeResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    get:
      tags: [Graph Service]
      summary: List nodes
      description: List nodes with optional mission_id/type filters.
      operationId: listNodes
      parameters:
        - name: mission_id
          in: query
          required: false
          schema:
            type: string
        - name: type
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/NodeType'
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 1000
      responses:
        '200':
          description: Node list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NodeQueryResponse'
...
  /api/v1/edges:
    post:
      tags: [Graph Service]
      summary: Create edge between nodes
      operationId: createEdge
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EdgeCreate'
      responses:
        '201':
          description: Edge created
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                  edge:
                    $ref: '#/components/schemas/EdgeResponse'
...
  /graphql:
    post:
      tags: [BFF Gateway]
      summary: GraphQL endpoint
      description: |
        GraphQL endpoint supporting queries, mutations, and subscriptions.

        ## Queries
        - `mission(id: String!)`: Get mission by ID
        - `missions(limit: Int, offset: Int)`: List missions
        - `nodes(mission_id: String!, filter: NodeFilter, limit: Int)`: Get nodes
        - `edges(mission_id: String!)`: Get edges
        - `graph_stats(mission_id: String!)`: Get graph statistics
        - `attack_paths(mission_id: String!, top: Int)`: Get attack paths
        - `workflow_nodes(mission_id: String!, types: [NodeType])`: Get workflow nodes
        - `workflow_layout(mission_id: String!)`: Get workflow layout

        ## Mutations
        - `start_mission(input: MissionInput!)`: Start new mission
        - `cancel_mission(id: String!)`: Cancel running mission
        - `save_workflow_layout(...)`: Save workflow layout

        ## Subscriptions
        - `graph_events(run_id: String!)`: Real-time graph events
        - `logs(run_id: String!)`: Real-time mission logs
      operationId: graphql
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GraphQLRequest'
      responses:
        '200':
          description: GraphQL response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GraphQLResponse'
...
  /api/v1/plan:
    post:
      tags: [Planner]
      summary: Generate attack plan
      operationId: generatePlan
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PlanRequest'
      responses:
        '200':
          description: Attack plan
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlanResponse'
  /api/v1/plan/{mission_id}:
    get:
      tags: [Planner]
      summary: Get attack plan for mission
      operationId: getPlan
      parameters:
        - name: mission_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Attack paths
          content:
            application/json:
              schema:
                type: object
                properties:
                  mission_id:
                    type: string
                  attack_paths:
                    type: array
                    items:
                      $ref: '#/components/schemas/AttackPath'

  # ============================================
  # ENDPOINT INTEL (Port 8004)
  # ============================================
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        detail:
          type: string
        status_code:
          type: integer

    # ============================================
    # ENUMS
    # ============================================
    NodeType:
      type: string
      enum:
        # Asset nodes
        - DOMAIN
        - SUBDOMAIN
        - HTTP_SERVICE
        - ENDPOINT
        - PARAMETER
        - HYPOTHESIS
        - VULNERABILITY
        - ATTACK_PATH
        - IP_ADDRESS
        - DNS_RECORD
        - ASN
        - ORG
        # Workflow nodes
        - AGENT_RUN
        - TOOL_CALL
        - LLM_REASONING
    EdgeType:
      type: string
      enum:
        # Asset relationships
        - HAS_SUBDOMAIN
        - RESOLVES_TO
        - SERVES
        - EXPOSES_HTTP
        - EXPOSES_ENDPOINT
        - HAS_PARAM
        - HAS_HYPOTHESIS
        - HAS_VULNERABILITY
        - TARGETS
        # Workflow relationships
        - TRIGGERS
        - USES_TOOL
        - PRODUCES
        - REFINES
        - LINKS_TO
...
    NodeCreate:
      type: object
      required:
        - id
        - type
        - mission_id
      properties:
        id:
          type: string
        type:
          $ref: '#/components/schemas/NodeType'
        mission_id:
          type: string
        properties:
          type: object
          additionalProperties: true
...
    NodeResponse:
      type: object
      properties:
        id:
          type: string
        type:
          $ref: '#/components/schemas/NodeType'
        mission_id:
          type: string
        properties:
          type: object
          additionalProperties: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    NodeQueryResponse:
      type: object
      properties:
        nodes:
          type: array
          items:
            $ref: '#/components/schemas/NodeResponse'
        total:
          type: integer
        limit:
          type: integer
        offset:
          type: integer

    # ============================================
    # EDGE SCHEMAS
    # ============================================
    EdgeCreate:
      type: object
      required:
        - from_node
        - to_node
        - edge_type
        - mission_id
      properties:
        from_node:
          type: string
        to_node:
          type: string
        edge_type:
          $ref: '#/components/schemas/EdgeType'
        mission_id:
          type: string
        properties:
          type: object
          additionalProperties: true
    EdgeResponse:
      type: object
      properties:
        id:
          type: string
        from_node:
          type: string
        to_node:
          type: string
        edge_type:
          $ref: '#/components/schemas/EdgeType'
        mission_id:
          type: string
        properties:
          type: object
          additionalProperties: true
        created_at:
          type: string
          format: date-time

    # ============================================
    # GRAPH SCHEMAS
    # ============================================
...
    GraphQLRequest:
      type: object
      required:
        - query
      properties:
        query:
          type: string
          description: GraphQL query string
        variables:
          type: object
          additionalProperties: true
        operationName:
          type: string
    GraphQLResponse:
      type: object
      properties:
        data:
          type: object
          additionalProperties: true
        errors:
          type: array
          items:
            type: object
            properties:
              message:
                type: string
              locations:
                type: array
                items:
                  type: object
              path:
                type: array
                items:
                  type: string

    # ============================================
    # PHASE EXECUTION SCHEMAS
    # ============================================
...
    PlanRequest:
      type: object
      required:
        - mission_id
      properties:
        mission_id:
          type: string
        top_k:
          type: integer
          default: 5
    PlanResponse:
      type: object
      properties:
        mission_id:
          type: string
        paths:
          type: array
          items:
            $ref: '#/components/schemas/AttackPath'
        computed_at:
          type: string
          format: date-time
    AttackPath:
      type: object
      properties:
        target:
          type: string
        score:
          type: integer
          minimum: 0
          maximum: 100
        actions:
          type: array
          items:
            type: string
        reasons:
          type: array
          items:
            type: string